requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.0.2
python-dotenv==1.0.0
PyMySQL==1.1.1
python-socketio==5.9.0
requests==2.32.3
kivy==2.3.1
